+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added `all` to repository.
 - Added `PersistentStates#mark_as_dirty`.
 - Added `Entity#references`.
-- Implemented composable entities!
+- Implemented nested composable entities!
 
 ## 0.40.0 [2013-05-21]